requirements.txt
# Cloud Resource Allocation – OpenEnv Environment
# ================================================

# OpenEnv core (server + client framework)
openenv-core

# Server dependencies
fastapi
uvicorn[standard]

# Environment physics
numpy

# Gradio demo (used when MODE=demo)
gradio>=4.0.0
matplotlib>=3.7.0

# DQN agent (optional, for Gradio demo mode)
# Using CPU-only torch to keep image small
torch --index-url https://download.pytorch.org/whl/cpu

# Inference script dependencies
openai